重构数据库配置为多数据源模式，添加MySQL支持，实现实体模块集成和用户查询接口

+++ daoyi-cloud-entity/src/entity/system/system_users.rs
@@ -4,7 +4,7 @@ use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(schema_name = "daoyi-system", table_name = "system_users")]
+#[sea_orm(table_name = "system_users")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
     #[sea_orm(primary_key)]
@@ -17,9 +17,9 @@ pub struct Model {
     pub post_ids: Option<String>,
     pub email: Option<String>,
     pub mobile: Option<String>,
-    pub sex: Option<i16>,
+    pub sex: Option<i8>,
     pub avatar: Option<String>,
-    pub status: i16,
+    pub status: i8,
     pub login_ip: Option<String>,
     pub login_date: Option<DateTime>,
     pub creator: Option<String>,
